Added pylint and test tasks
This commit includes a new setup.py 'ci' command, which runs pylint
and a couple of new smoke tests which run the application and client
generation commands. The commit also includes a few cosmetic code
updates to conform to pylint rules.

This commit also expands the Travis test matrix to include 3.4, 3.5, and
3.6 and removes tasks from the Travis job which are unrelated to running
the pylint and test tasks.