Avoid trying to copy directories under _config
Previously, some logic was added to copy files from the app/_config
directory to the target config directory when not present at startup. If
the file being copied, however, was a directory - for example, a
__pycache__ directory for Python 3 - the copy logic would throw an error
and abort. This commit changes the logic to avoid the error by ignoring
any directories that might exist under the _config directory.