Replace usages of builtin configparser NoOptionError
Previously, the template parsing logic would re-raise a NoOptionError
when the underlying ConfigParser was unable to find the requested
option. Because this class is in a different location in Python 2 vs.
Python 3, this required some version-specific code try to treat the
error the same across versions. This commit avoids the use of the
builtin NoOptionError in order to avoid this conditional logic. Various
locations which raised an exception when an option could not be found
now raise a new internal NoOptionError exception, which would be in the
same module location across Python versions.