Strip leading copyright word from conf.py copyright key
For the copyright option in the app and client config templates, users
typically insert the word "Copyright" at the front. Where this is used
in the generation of the conf.py file, however, the Sphinx tool inserts
its own "Copyright" word in generated HTML docs, resulting in
"Copyright Copyright" appearing at the bottom of each HTML page. For
this commit, the word Copyright is stripped from the front of the
copyright keyword in the generated conf.py file, avoiding the appearance
of double "Copyright" words in the generated HTML pages.