Coerce values read from config to unicode strings
Previously, when values were read from the config, they would be coerced
to unicode strings if the configparser package was installed on Python 2
but would be bytes (str) if not. Leaving the value as bytes would result
in exceptions being thrown by consumers expecting the values to be
unicode strings. In this commit, the values are coerced to unicode
strings whether or not the configparser package is installed, avoiding
the exceptions which were previously thrown.